requirements_setup.txt
# 🚀 Niche Compass Setup Scripts Requirements
# Install with: pip install -r requirements_setup.txt

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Database dependencies
pymongo>=4.6.0
tinydb>=4.8.0

# Azure AI services
azure-cognitiveservices-vision-computervision>=0.9.0
azure-ai-textanalytics>=5.3.0
openai>=1.0.0

# Flask and web framework
Flask>=2.3.0
Flask-CORS>=4.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Image processing (for visual analysis)
Pillow>=10.0.0

# JSON and configuration
PyYAML>=6.0.1

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and monitoring
colorama>=0.4.6
rich>=13.5.0

# HTTP and API testing
httpx>=0.24.0

# Date and time utilities
python-dateutil>=2.8.2

# Environment and configuration management
python-decouple>=3.8

# Optional: For advanced features
# opencv-python>=4.8.0  # Uncomment if you need image processing
# scikit-learn>=1.3.0   # Uncomment if you need ML features
